Adding InlineAdaptiveBannerExample.
PiperOrigin-RevId: 604577685
