+++ .github/workflows/build.yml
@@ -19,7 +19,7 @@ jobs:
       matrix:
         regex:
         - "java/advanced/APIDemo"
-        - "java/advanced/BannerRecyclerViewExample"
+        - "java/advanced/InlineAdaptiveBannerExample"
         - "java/advanced/RewardedSSVExample"
         - "kotlin/advanced/APIDemo"
         - "kotlin/advanced/InlineAdaptiveBannerExample"
