--- java/advanced/BannerRecyclerViewExample/app/build.gradle
@@ -1,27 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    namespace 'com.google.android.gms.example.bannerrecyclerviewexample'
-    compileSdkVersion 34
-    defaultConfig {
-        applicationId "com.google.android.gms.example.bannerrecyclerviewexample"
-        minSdkVersion 19
-        multiDexEnabled true
-        targetSdkVersion 34
-        versionCode 1
-        versionName "1.0"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'androidx.recyclerview:recyclerview:1.0.0'
-    implementation 'com.google.android.gms:play-services-ads:22.6.0'
-}
