--- java/advanced/BannerRecyclerViewExample/app/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">Banner RecyclerView Example</string>
-    <string name="ad_unit_id">ca-app-pub-3940256099942544/1072772517</string>
-</resources>
