+++ java/advanced/InlineAdaptiveBannerExample/app/.gitignore
