+++ java/advanced/InlineAdaptiveBannerExample/app/build.gradle
@@ -1,0 +1,27 @@
+apply plugin: 'com.android.application'
+
+android {
+    namespace 'com.google.android.gms.example.inlineadaptivebannerexample'
+    compileSdk 34
+    defaultConfig {
+        applicationId "com.google.android.gms.example.inlineadaptivebannerexample"
+        minSdk 21
+        targetSdkVersion 34
+        versionCode 1
+        versionName "1.0"
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+
+dependencies {
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.3.2'
+    implementation 'com.google.android.gms:play-services-ads:22.6.0'
+    implementation 'androidx.window:window:1.2.0'
+}
