+++ java/advanced/InlineAdaptiveBannerExample/app/proguard-rules.pro
