+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/layout/banner_ad_container.xml
