+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/layout/menu_item_container.xml
