+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/raw/menu_items_json.json
