+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/values-w820dp/dimens.xml
