+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/values/colors.xml
