+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/values/dimens.xml
