+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name" translatable="false">Inline Adaptive Banner RecyclerView Example</string>
+</resources>
