+++ java/advanced/InlineAdaptiveBannerExample/app/src/main/res/values/styles.xml
