+++ java/advanced/InlineAdaptiveBannerExample/gradle.properties
