+++ java/advanced/InlineAdaptiveBannerExample/gradle/wrapper/gradle-wrapper.properties
