+++ java/advanced/InlineAdaptiveBannerExample/settings.gradle
